% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddIfUnique.R
\name{AddIfUnique}
\alias{AddIfUnique}
\title{Appending functions for accumulating weight vectors}
\usage{
AddIfUnique(df, row)
}
\arguments{
\item{lambda}{Weight vector with two numerical components (or only the first two will be checked)}
}
\value{
The data frame, possibly with appended new row
}
\description{
Utility function for exact decomposition algorithm.
Append the row to data frame only if it does not already exist within it.
}
\examples{
Lambda <- weight_set(0.5)
weight <- c(0.1,0,0.9, -0.4, 0)
Lambda <- AddIfUnique(Lambda, weight)
}
