% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition_intersections.R
\name{decomposition_intersections}
\alias{decomposition_intersections}
\title{Compute intersections from line segments for exact rank aggregation decomposition}
\usage{
decomposition_intersections(Linedf, show_bar = FALSE)
}
\arguments{
\item{Linedf}{Data frame \link{decomposition_linesegs} where each row describes a line segment.}

\item{show_bar}{Boolean for a progress bar and time data printed in console}
}
\value{
A data frame of two columns (x and y) for all intersection points in
weight set decomposition, including those on the boundary.
}
\description{
Every pair of computed line segments are checked for an intersection within the
boundary of Lambda.
Step 2 of \link{rank_decomposition_exact}
}
\examples{
rankdf <- data.frame('rank1'=c(1,2,3,4,5), 
           'rank2'=c(2,3,1,5,4), 'rank3'=c(3,1,5,4,2))
Linedf = decomposition_linesegs(rankdf)
IntersectPts = decomposition_intersections(Linedf)
}
