% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition_linesegs.R
\name{decomposition_linesegs}
\alias{decomposition_linesegs}
\title{Compute line segments for exact rank aggregation decomposition}
\usage{
decomposition_linesegs(input_data, show_bar = FALSE)
}
\arguments{
\item{input_data}{Data frame containing 3 columns of ranks/ratings per item.
Each row represents one item.}

\item{show_bar}{Boolean for a progress bar and time data printed in console}
}
\value{
A data frame of labeled line segments contained within Lambda.
Lines are labeled by two boundary points given by x & y: (bp1x, bp1y) and (bp2x, bp2y).
The slope and y-intercept are given in two columns, as well as the item ids
used to compute the line.
}
\description{
Every pair of items is used to compute a line. The intersections of the line with
boundary of Lambda (if there are any) are called boundary points.
Computes all boundary points and returns.
Step 1 of \link{rank_decomposition_exact}
}
\examples{
rankdf <- data.frame('rank1'=c(1,2,3,4,5), 
           'rank2'=c(2,3,1,5,4), 'rank3'=c(3,1,5,4,2))
Linedf = decomposition_linesegs(rankdf)
Linedf = decomposition_linesegs(rankdf, show_bar=TRUE)
}
