% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equi_transform.R
\name{equi_transform}
\alias{equi_transform}
\title{Transform weights to the equilateral triangle representation}
\usage{
equi_transform(Lambda)
}
\arguments{
\item{Lamdba}{Data frame with column names lambda1 and lambda2, representing weights for the
right triangle representation}
}
\value{
An extended data frame with extra columns named equilambda1 and equilambda2
}
\description{
Appends columns to the data frame from \code{\link[=weight_set]{weight_set()}} which transforms the right triangle representation
(lambda1 and lambda2) to the equilateral triangle representation (equilambda1 and equilambda2).
Note equilambda1 is centered at 0, so it is expected to have negative values.
These weights are recommended for less biased visualizations.
}
\examples{
Lambda1 <- weight_set()
Lambda2 <- equi_transform(Lambda1)
}
