% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_rank.R
\name{flex_rank}
\alias{flex_rank}
\title{Flexible ranking function}
\usage{
flex_rank(vec, ties = FALSE)
}
\arguments{
\item{vec}{Vector of scores/ratings to be ranked where smaller values are better}

\item{ties}{Boolean to indicate whether the ranking should allow for ties (TRUE) or not (FALSE)}
}
\value{
A vector of the same length in which the ith value is the rank of the ith item
}
\description{
Provides two ways of ranking input vector of scores/ratings. The default method does not
allow for ties, so in the case of a tie, it is broken by order of indices. The alternative
method allows for ties, and it is computed by the number of smaller values.
}
\examples{
risk <- c(1.0,  2.0,  3.1,  2.0,  1.7)
flex_rank(risk)
flex_rank(risk,ties=TRUE)
}
