% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InLambda.R
\name{InLambda}
\alias{InLambda}
\title{In Lambda boolean for weight vectors}
\usage{
InLambda(lambda)
}
\arguments{
\item{lambda}{Weight vector with two numerical components (or only the first two will be checked)}
}
\value{
True (if yes) or False (if no)
}
\description{
Utility function for exact decomposition algorithm.
Returns true if the weight vector satisfies the (right-triangle) representation of weight set.
2 dimensional: x + y <= 1 and x,y >= 0.
3 dimensional: x + y + z == 1 and x,y,z >= 0.
}
\examples{
InLambda(c(0.5,0.5))
InLambda(c(0.6,0.5))
InLambda(c(0.5,0.5,0.5))
}
