% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex_knapsack.R
\name{lex_knapsack}
\alias{lex_knapsack}
\title{Lexicographic solution to knapsack problem}
\usage{
lex_knapsack(rank.label, costs, budget)
}
\arguments{
\item{rank.label}{String describing the ranked position per item}

\item{costs}{Array of costs per item}

\item{budget}{Total budget which cannot be exceeded}
}
\value{
A string describing the items included in the lexicographic solution in order of index
}
\description{
Provides the set of items for the lexicographic (or greedy) solution for a single budget constraint.
Items are choosen by following the given rank vector: an item is only packed if it fits in current budget.
}
\examples{
rank.label = '4.6.3.1.1.4'
costs <- c(1.0,  0.5,  2.1,  4.0,  4.7, 1.5)
lex_knapsack(rank.label,costs,budget=6)
lex_knapsack(rank.label,costs,budget=7)
lex_knapsack(rank.label,costs,budget=8)
}
