% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram_exact.R
\name{plot_histogram_exact}
\alias{plot_histogram_exact}
\title{Plot a histogram of indifference region areas from exact decomposition}
\usage{
plot_histogram_exact(
  IR_hull,
  orientation = "horizontal",
  min_threshold = 0,
  axis_labels = TRUE,
  leg.pos = "none"
)
}
\arguments{
\item{IR_hull}{Data frame from \link{rank_decomposition_exact} containing weight vectors and labels for each indifference region.
Must include column name Rank.Label for the string/factor identifier of the rank aggregation.}

\item{orientation}{'vertical' or 'horizontal' for the type of bar graph.}

\item{min_threshold}{Number between 0 and 100 which specifies a minimum threshold of area (percentage).
Any regions with area less than the threshold is omitted from the histogram.}

\item{axis_labels}{Specify whether axis should be labeled by the complete ranked list}

\item{leg.pos}{Input for ggplot legend positioning, e.g. 'bottom', 'right', or c(0.8,0.8)}
}
\value{
list containing: [\link{1}] A ggplot structure which can be plotted directly and
[\link{2}] the data frame with estimated areas
}
\description{
Using the labeled rank aggregations from the exact decomposition, \link{rank_decomposition_exact},
plot the area of each
indifference region as a percentage of the entire weight set.
}
\examples{
Lambda <- weight_set(stepsize=0.1)
rankdf <- data.frame('rank1'=c(1,2,3,4,5), 
           'rank2'=c(2,3,1,5,4), 'rank3'=c(3,1,5,4,2))
Lambda <- Lambda <- rank_aggregation_grid(3,Lambda,rankdf)
inputlist <- rank_decomposition_exact(rankdf,Lambda)
IR_hull <- inputlist[[4]]
outlist <- plot_histogram_exact(IR_hull)
outlist[[1]] # for ggplot
outlist[[2]] # for data frame
plot_histogram_exact(IR_hull,orientation='vertical',min_threshold=1)[[1]]
plot_histogram_exact(IR_hull,min_threshold=1,axis_labels=FALSE,leg.pos='bottom')[[1]]
}
