% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_position_grid.R
\name{plot_position_grid}
\alias{plot_position_grid}
\title{Plot one position region for the grid search}
\usage{
plot_position_grid(
  Lambda,
  item,
  position = "all",
  triangle = "equilateral",
  bias_axes = TRUE,
  annotations = TRUE,
  plotly_text = FALSE,
  leg.pos = "none"
)
}
\arguments{
\item{Lambda}{Data frame containing weight vectors and labels for weighted rank aggregations.
Must include column names lambda1/2 for right triangle representation or
equilambda1/2 for equilateral triangle representation.
Must include column name Label for the string/factor identifier of the rank aggregation.}

\item{item}{The index for which item to represent.}

\item{position}{Either string 'all' to illustrate all positions or a specific numeric}

\item{triangle}{Specify whether the right triangle representation ("right") or the
equilateral triangle representation ("equilateral"). Note that the equilateral triangle
representation is recommended as it is unbiased, and the right triangle representation is
biased.}

\item{bias_axes}{Boolean for whether to include bias axes in the graph}

\item{annotations}{Boolean for whether to include annotations in the graph}

\item{plotly_text}{Boolean for whether to include text which is useful for plotly conversion.
Note that a warning will occur about an "unknown aesthetics" (text); this is a
feature which is intended to be used by the ggplotly interactive library. The warning is not
fixable, and should be ignored. For ggplotly visual tool, use
ggplotly(g,tooltip = 'text') to access these aesthetics as appropriate labels.}

\item{leg.pos}{Input for ggplot legend positioning, e.g. 'bottom', 'right', or c(0.8,0.8)}
}
\value{
A ggplot structure which can be plotted directly.
}
\description{
Choose one of the items to show the item's ranked position for all weights in the weight set grid.
}
\examples{
Lambda <- weight_set()
metrics <- data.frame('cost'=c(10,20,30,40), 'time'=c(5.9, 3.3, 2.5, 4.1), 'risk'=c(1,4,3,2))
Lambda <- rank_aggregation_grid(3,Lambda,metrics)
plot_position_grid(Lambda,item=1)
g <- plot_position_grid(Lambda,item=1,leg.pos='bottom')
g
plot_position_grid(Lambda,item=1,position=1)
plot_position_grid(Lambda,item=1,position='all',triangle='right',leg.pos='bottom')
}
