% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_labels.R
\name{randomize_labels}
\alias{randomize_labels}
\title{Randomize the order of levels of a factor for better color separation in graphs}
\usage{
randomize_labels(label)
}
\arguments{
\item{label}{Column or array of labels, e.g., Rank.Label or Item.Label}
}
\value{
Array of same labels but as a factor with the levels randomly shuffled
}
\description{
Randomly reorders the levels of some factor's labels. This helps when graphing because neighboring
indifference regions tend to have the same color. By randomizing the order, the colors are
more randomized and distinguishable. Recommended use for Lambda$Rank.Label or Lambda$TopAlpha.
}
\examples{
Lambda <- weight_set(0.01)
metrics <- data.frame('cost'=c(10,20,30,40), 'time'=c(5.9, 3.3, 2.5, 4.1), 'risk'=c(1,4,3,2))
Lambda <- rank_aggregation_grid(Lambda,metrics)
Lambda$Rank.Label <- randomize_labels(Lambda$Rank.Label)
}
