% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_budget_grid.R
\name{rank_budget_grid}
\alias{rank_budget_grid}
\title{Budgeted ranking over the grid of weights}
\usage{
rank_budget_grid(Lambda, costs, budget, metrics)
}
\arguments{
\item{Lambda}{Data frame containing columns lambda1, lambda2, lambda3.
Runs faster if it already contains column Rank.Label from \link{rank_aggregation_grid}
(factor/string of ordered items separated by periods)}

\item{costs}{Array of costs per item}

\item{budget}{Maximum budget which cannot be exceeded}

\item{metrics}{Data frame for 3 metrics, one metric per column, one row per item, smaller values are better.
Only required if Rank.Label column is missing from Lambda.}
}
\value{
An extended data frame with column for Budget, which are factors (of strings) that list
the solution items (ordered numerically not by rank)
}
\description{
Given costs per item and a maximum budget, give the lexicographic or greedy knapsack solution per aggregate rank.
The order of the items in the knapsack solution are ignored. All rank aggregations are
relabeled according to the items in the solution. Note that if Rank.Label is not provided in the
input data frame, then they will first be computed (without ties).
}
\examples{
Lambda <- weight_set()
metrics <- data.frame('risk1'=c(1,2,3,4,5), 
           'risk2'=c(2,3,1,5,4), 'risk3'=c(3,1,5,4,2))
costs <- c(4.1, 4.4, 2.2, 5.5, 3.6 )
Lambda <- rank_budget_grid(Lambda, costs=costs, budget=10, metrics=metrics)
plot_aggregation_grid(Lambda,by='Budget',leg.pos='bottom')
}
