% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_topalpha_grid.R
\name{rank_topalpha_grid}
\alias{rank_topalpha_grid}
\title{Top-alpha ranking over the grid of weights}
\usage{
rank_topalpha_grid(Lambda, alpha = 5, metrics)
}
\arguments{
\item{Lambda}{Data frame containing columns lambda1, lambda2, lambda3.
Runs faster if it already contains column Rank.Label from \link{rank_aggregation_grid}
(factor/string of ordered items separated by periods)}

\item{alpha}{The number of top-ranked positions of interest. Must be at least 2.}

\item{metrics}{Data frame for 3 metrics, one metric per column, one row per item, smaller values are better.
Only required if Rank.Label column is missing from Lambda.}
}
\value{
An extended data frame with column for TopAlpha, which are factors (of strings) that list
the top ranked items (ordered numerically not by rank)
}
\description{
Alpha is the top number of positions of interest, for instance the top-5 or top-10 items in
the ranked list. The order of the items in the top alpha are ignored. All rank aggregations are
relabeled according to their top-alpha items. Note that if Rank.Label is not provided in the
input data frame, then they will first be computed (without ties).
}
\examples{
Lambda <- weight_set(stepsize=0.1)
metrics <- data.frame('cost'=c(10,20,30,40,50), 
           'time'=c(5.9, 3.3, 2.5, 4.1, 1.8), 'risk'=c(1,4,3,2,5))
Lambda <- rank_topalpha_grid(Lambda, alpha=3, metrics=metrics)
plot_aggregation_grid(Lambda,by='TopAlpha')
}
