% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_set.R
\name{weight_set}
\alias{weight_set}
\title{Initialize the data frame for weights in weight set}
\usage{
weight_set(stepsize = 0.01, equilateral = TRUE)
}
\arguments{
\item{stepsize}{The step size for the grid of equally spaced points.
Smaller values will increase precision but requires more time for computation.
Weights are between 0 and 1 with sum of 1.}

\item{equilateral}{Binary to indicate whether or not to include the equilateral transformation
as columns equilambda1 and equilambda2}
}
\value{
A data frame where each row represents one grid point in the weight set.
}
\description{
Returns a data structure to organize weights equally spaced in the weight set.
For 3 input scores/metrics, the weights are labeled by column as lambda1, lambda2, lambda3,
where lambda1 and lambda2 correspond to the right triangle representation.
Every row represents one weight vector. All weights are nonnegative.
All weight vectors sum to one. These weights are recommended to be used for all computations.
These weights may also be used for visualization (lambda1 and lambda2 only), but note that it
is a biased representation since the distance between vertices of the triangle are not uniform.
}
\examples{
Lambda <- weight_set() 
}
